+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - The image regular expression is now a little bit less restrictive, allowing for some spaces before and after the image.
 - Fixed a small bug during checking and unchecking of task list items that would prevent the updating of the underlying Markdown text.
 - When you enter an ID and choose the ID from the popup list, the filename belonging to that ID will also be inserted after the ID.
+- You can now filter the tags in the tag cloud.
 
 ## Under the Hood